[WEB-1754] Put error information directly in log
We are seeing these errors getting logged and yet the error information
we are trying to embed in the log metadata is not there. This makes
working out what really happened a lot harder.

So, here we instead try to get the error information into a the
log string itself so that we can see what's happening in our logs.

Test plan:

We don't have a great story right now for testing our failure modes.
However, we can test this by sending a bad request that requests URLs
that we know don't exist, or by sending a request that uses JS that we
know will blow up when executed.

To do the second part, we need to serve a file that is broken. We can
coerce a webapp to do this.