[WEB-1753] Add abort to Promise prototype
This makes it so that all our promises are patched to have an abort
call by modifying the Promise prototype. The aim is to make sure that
when JSDOM closes, there is an abort function for it to call, working
around its bug where it assumes all fetched promises are abortable this way.

This also takes the `CustomResourceLoader` code and gives it its own
file and tests. This helps validate a core piece of our architecture
as well as verify that our patching is working.