WEB-1980: Clear unused files from fetch cache to reduce OOM errors (#58)
This ups the cache expiration to 24 hours (as the files themselves shouldn't be changing) and then uses the new logic for clearing the cache of unused files every 15 minutes. I think we'll want to test this in practice to see if there's a duration that makes more sense but I think this is a good starting point.

Related PR: https://github.com/Khan/cache-service-cache-module/pull/1

Issue: https://khanacademy.atlassian.net/browse/WEB-1980

Test plan:
`yarn test` passed. I think the rest we'll want to test in production to see how it works!