WEB-1978: Use .pruneKey() to have non-EN JS files use EN files (#59)
This uses the new logic added here:
https://github.com/Khan/superagent-cache-plugin/pull/2

To make it so that we can modify the cache key of superagent-cache-plugin, making it so that the URI is now always "en", regardless of the locale. This is safe to do as the filenames are actually unique: they contain unique hashes in them (which we change if they've been translated). If a file hasn't had an translation done to them then the file is essentially the same as the en version so we can use them interchangably, allowing us to cut down on the number of files we need to download and store.

Issue: https://khanacademy.atlassian.net/browse/WEB-1978

Test plan:
I ran `yarn test` and it passed.